+++ lib/features/reports/domain/model/uncategories_row_data.dart
@@ -1,4 +1,4 @@
-import '../../../features/categories/domain/model/category.dart';
+import '../../../categories/domain/model/category.dart';
 
 class UncategorisedRowData {
   Category category;
