+++ lib/features/reports/domain/usecase/update_categories_from_data_usecase.dart
@@ -1,4 +1,4 @@
-import '../../../../shared/domain/model/uncategories_row_data.dart';
+import '../model/uncategories_row_data.dart';
 import '../../../categories/domain/model/category.dart';
 
 class UpdateCategoriesFromRowData {
