+++ lib/features/reports/presentation/ui/uncategorised_item_row.dart
@@ -2,7 +2,7 @@ import 'package:expense_categoriser/features/reports/presentation/ui/selectable_
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../shared/domain/model/uncategories_row_data.dart';
+import '../../domain/model/uncategories_row_data.dart';
 import '../../domain/model/selectable_word_item.dart';
 import '../../domain/model/transaction.dart';
 import '../../../categories/domain/model/category.dart';
