+++ lib/features/reports/presentation/view/report_screen.dart
@@ -5,7 +5,7 @@ import 'package:expense_categoriser/features/reports/presentation/viewmodel/repo
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../shared/domain/model/uncategories_row_data.dart';
+import '../../domain/model/uncategories_row_data.dart';
 import '../../domain/model/transaction.dart';
 import '../../../categories/presentaion/viewmodel/categories_viewmodel.dart';
 import '../ui/uncategorised_item_row.dart';
