+++ lib/shared/presentation/ui/app_menu.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../constants.dart';
+import '../../constants.dart';
 
 final selectedPageNameProvider = StateProvider<String>((ref) {
   return pages.keys.first;
